Move unit of work interface to application layer
