+++ source/Processing.Infrastructure/BusinessRequestProcessing/Pipeline/UnitOfWorkBehaviour.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
+using Processing.Application.Common;
 using Processing.Domain.SeedWork;
 using Processing.Infrastructure.Configuration.DataAccess;
